Task 1 and 4 completed, Task 2 completed but not fully tested

+++ 6_OOPPrinciples_Debugging/src/BankManagement/Bank.java
@@ -8,7 +8,7 @@ public class Bank {
     Account[] accounts = new Account[200];
     int accountIndex = 0;
 
-    void openAccount(String name, int govid, String username, String password) {
+    void openAccount(String name, long govid, String username, String password) {
         account = new Account(name, govid, username, password);
         accounts[accountIndex] = account;
         if (accountIndex == 200) {
